+++ src/backend/resolvers/mutations/organization/meta/index.js
@@ -1,0 +1,5 @@
+const upsertOrganizationMeta = require('./upsert')
+
+module.exports = {
+  upsertOrganizationMeta,
+}