+++ src/backend/resolvers/queries/organizations/meta.js
@@ -1,0 +1,14 @@
+const { ObjectId } = require('mongodb')
+
+module.exports = (
+  parent,
+  { _ids },
+  { pulseCoreDb },
+) => pulseCoreDb.collection('organizations.meta')
+  .find({
+    accountId: {
+      $in: _ids.map(_id => ObjectId(_id)),
+    }
+  })
+  .toArray()
+